services/user_service/requirements.txt
fastapi
uvicorn
psycopg2-binary
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-jaeger